refactor: update ClaimDetails container
