+++ src/containers/ClaimDetails/LoadingState.tsx
