+++ src/containers/ClaimDetails/RenderClaimDetails.tsx
